Vizualizer moved to a seperate notebook.
Altough it almost works perfectly it will not be included in the handin. This is partly because it introduces so much extra complexity unrelated to the machine learning models but mostly because the final visualization is quite hard to interpret when you don't have a rubiks cube in front of you. A simpler visualisation with only three sides from the camera perspective would suit this task much better.